Description
This code takes you from a list of genotypes to a table of pairwise Fst values among all populations. The starting file used here is a text file with one individual per row, no individual code (including this gave an error), genotypes listed with a / (e.g., A/a), and a column of populations. The starting file (MHC.loci.txt) has been uploaded to this directory.