Create plot-haplotype-network-in-r
This script takes you from a fasta file of haplotypes to a haplotype network plot. In here I also mess around with color schemes a little via RColorBrewer. I couldn't get the pies to not overlap in r (the only suggestion I found was to change the scale of the whole plot, which makes the branches get too long) so I saved the haplotype network plot as a .svg and edited it in InkScape.

The input file (haplotypes_complete_seqs.fasta) has been uploaded to this directory.